changed: sonames to match ffmpeg library names on linux
We rename the ffmpeg libraries to custom names, but the SONAME and
NEEDED tags in the libraries are not modified. This confuses some
dependency generators into thinking they require the system FFmpeg.
Change sonames to match file names to satisfy them. As a bonus, this
prevents the dynamic linker from linking in system FFmpeg libraries for
the dependencies in case the internal dependency was not already loaded
by XBMC in case of a bug.